fix make scripts, add service creation and return created trader id

+++ Makefile
@@ -1,9 +1,9 @@
 compile:
-	rm dist/trader
+	if [ -e .dist/trader ];	then rm dist/trader; fi;
 	GOOS=linux go build -o dist/trader .
 
 remove-build-docker:
-	docker image rm coiner/trader-provisioner:latest
-	docker build -t coiner/trader-provisioner:latest .
+	if [[ "$(docker images -q coiner/trader-provisioner:0.0.1 2> /dev/null)" != "" ]]; then docker image rm coiner/trader-provisioner:0.0.1; fi;
+	docker build -t coiner/trader-provisioner:0.0.1 .
 
 dockerize: compile remove-build-docker
